Create styling for home page link
Initially created an anchor tag with a following paragraph tag for the link text.

Removing the 'p' tag allowed me to reduce the amount of css code need to make the desired outcome.

This is because the body tag already has a background-color set.

Changing this with the class selector a.home-page-link would not overwrite the body tag selector.